backend/requirements.txt
# FastAPI and ASGI server
fastapi
uvicorn[standard]
python-multipart

# Database
sqlalchemy
psycopg2-binary
alembic

# Authentication
python-jose[cryptography]
passlib[bcrypt]

# Face Recognition and Image Processing
face-recognition
opencv-python
Pillow
numpy

# Data validation and settings
pydantic
pydantic-settings

# HTTP requests
httpx
requests

# Environment and configuration
python-dotenv

# CORS
fastapi-cors

# Date and time
python-dateutil

# Logging
loguru

# Testing
pytest
pytest-asyncio

# Development
black
flake8
isort
